+++ tests/framework/mail/BaseMessageTest.php
@@ -8,9 +8,7 @@
 namespace yiiunit\framework\mail;
 
 use Yii;
-use yii\mail\BaseMailer;
-use yii\mail\BaseMessage;
-use yii\mail\MessageInterface;
+use yiiunit\data\mail\TestMailer;
 use yiiunit\TestCase;
 
 /**
@@ -22,13 +20,13 @@ class BaseMessageTest extends TestCase
     {
         $this->mockApplication([
             'components' => [
-                'mailer' => $this->createTestEmailComponent(),
-            ],
+                'mailer' => $this->createTestEmailComponent()
+            ]
         ]);
     }
 
     /**
-     * @return Mailer test email component instance.
+     * @return TestMailer test email component instance.
      */
     protected function createTestEmailComponent()
     {
@@ -61,129 +59,4 @@ class BaseMessageTest extends TestCase
         $message = $mailer->compose();
         $this->assertEquals($message->toString(), '' . $message);
     }
-}
-
-/**
- * Test Mailer class
- */
-class TestMailer extends BaseMailer
-{
-    public $messageClass = 'yiiunit\framework\mail\TestMessage';
-    public $sentMessages = [];
-
-    protected function sendMessage($message)
-    {
-        $this->sentMessages[] = $message;
-    }
-}
-
-/**
- * Test Message class
- */
-class TestMessage extends BaseMessage
-{
-    public $text;
-    public $html;
-
-    public function getCharset()
-    {
-        return '';
-    }
-
-    public function setCharset($charset)
-    {
-    }
-
-    public function getFrom()
-    {
-        return '';
-    }
-
-    public function setFrom($from)
-    {
-    }
-
-    public function getReplyTo()
-    {
-        return '';
-    }
-
-    public function setReplyTo($replyTo)
-    {
-    }
-
-    public function getTo()
-    {
-        return '';
-    }
-
-    public function setTo($to)
-    {
-    }
-
-    public function getCc()
-    {
-        return '';
-    }
-
-    public function setCc($cc)
-    {
-    }
-
-    public function getBcc()
-    {
-        return '';
-    }
-
-    public function setBcc($bcc)
-    {
-    }
-
-    public function getSubject()
-    {
-        return '';
-    }
-
-    public function setSubject($subject)
-    {
-    }
-
-    public function setTextBody($text)
-    {
-        $this->text = $text;
-    }
-
-    public function setHtmlBody($html)
-    {
-        $this->html = $html;
-    }
-
-    public function attachContent($content, array $options = [])
-    {
-    }
-
-    public function attach($fileName, array $options = [])
-    {
-    }
-
-    public function embed($fileName, array $options = [])
-    {
-    }
-
-    public function embedContent($content, array $options = [])
-    {
-    }
-
-    public function toString()
-    {
-        return get_class($this);
-    }
-
-    public function addHeader($name, $value) {}
-
-    public function setHeader($name, $value) {}
-
-    public function getHeader($name) {}
-
-    public function setHeaders($headers) {}
 }